The Node Sever now actually serves the react app
Now instead of using  npm run in both the client and server, you can simply build  the client once. Then when running the server it will serve the built app

While it's  still only running on LocalHost this is closer to the functionality the program will have in final deployment.