Added Minimum Remaining Value Ordering
Added Minimum Remaining Value (MRV) Ordering this means that the nodes with the least values will be assigned first. This helps by creating a "fail fast" environment where if assignment is impossible it  reaches it faster. Also ends up helping efficiently overall slightly.

Also changes to the example were made so it is slightly more complex, it's based of the Australia coloring problem. Just google it, if you want to know more.